pyproject.toml
[project]
name = "cemp-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pyspark>=4.1.1",
]

[project.optional-dependencies]
data-gen = [
    "ulid-py>=1.1.0",
]
